# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateVendorRequest:
    properties:
      address:
        type: string
      description:
        type: string
      email:
        type: string
      image_url:
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
      phone_number:
        type: string
    required:
    - name
    - phone_number
    type: object
...
  dto.EmptySuccessResponseWrapper:
    properties:
      message:
        example: success
        type: string
      status:
        example: true
        type: boolean
    type: object
  dto.ErrorResponse:
    properties:
      error:
        example: invalid request
        type: string
      message:
        example: failed request
        type: string
      status:
        example: false
        type: boolean
    type: object
...
  dto.UpdateVendorRequest:
    properties:
      address:
        type: string
      description:
        type: string
      email:
        type: string
      image_url:
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
      phone_number:
        type: string
    required:
    - name
    - phone_number
    type: object
...
  dto.VendorResponse:
    properties:
      address:
        type: string
      description:
        type: string
      email:
        type: string
      id:
        type: string
      image_url:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
  dto.VendorResponseWrapper:
    properties:
      data:
        $ref: '#/definitions/dto.VendorResponse'
      message:
        example: success get vendor
        type: string
      status:
        example: true
        type: boolean
    type: object
  dto.VendorsResponseWrapper:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.VendorResponse'
        type: array
      message:
        example: success get vendors
        type: string
      meta: {}
      status:
        example: true
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@simponi.app
    name: Simponi Support
  description: |-
    REST API for Simponi - multi-store inventory and omnichannel management system.
    This API supports store management, product management, external product integration, vendor management, and role-based access control (RBAC).
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://simponi.app/terms/
  title: Simponi Backend API
  version: "1.0"
paths:
...
  /stores/{store_id}/vendors:
    get:
      consumes:
      - application/json
      description: 'Get paginated vendors in a store (Requires permission: GetVendors)'
      parameters:
      - description: Store ID (UUID)
        in: path
        name: store_id
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/dto.VendorsResponseWrapper'
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Store not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get vendors
      tags:
      - Vendors
    post:
      consumes:
      - application/json
      description: 'Create a new vendor inside a store (Requires permission: CreateVendor)'
      parameters:
      - description: Store ID (UUID)
        in: path
        name: store_id
        required: true
        type: string
      - description: Create vendor request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.CreateVendorRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Success
          schema:
            $ref: '#/definitions/dto.VendorResponseWrapper'
        "400":
          description: Invalid input / UUID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Store not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create vendor
      tags:
      - Vendors
  /stores/{store_id}/vendors/{vendor_id}:
    delete:
      consumes:
      - application/json
      description: 'Delete vendor from a store by vendor ID (Requires permission:
        DeleteVendorByID)'
      parameters:
      - description: Store ID (UUID)
        in: path
        name: store_id
        required: true
        type: string
      - description: Vendor ID (UUID)
        in: path
        name: vendor_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/dto.EmptySuccessResponseWrapper'
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Vendor not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete vendor
      tags:
      - Vendors
    get:
      consumes:
      - application/json
      description: 'Get detail vendor in a store by vendor ID (Requires permission:
        GetVendorByID)'
      parameters:
      - description: Store ID (UUID)
        in: path
        name: store_id
        required: true
        type: string
      - description: Vendor ID (UUID)
        in: path
        name: vendor_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/dto.VendorResponseWrapper'
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Vendor not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get vendor by ID
      tags:
      - Vendors
    put:
      consumes:
      - application/json
      description: 'Update vendor data in a store (Requires permission: UpdateVendor)'
      parameters:
      - description: Store ID (UUID)
        in: path
        name: store_id
        required: true
        type: string
      - description: Vendor ID (UUID)
        in: path
        name: vendor_id
        required: true
        type: string
      - description: Update vendor request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateVendorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/dto.VendorResponseWrapper'
        "400":
          description: Invalid UUID / Request Body
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Vendor not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update vendor
      tags:
      - Vendors
...
securityDefinitions:
  BearerAuth:
    description: 'Use JWT token with format: Bearer <your_token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...